new stress-strain model
The formula previously permitted epsilons in individual layers that
could exceed 1 (meaning a thinned-to-negative layer). And permitting
that meant that we never really compressed the hard layers (like bone).
I now remove layers that have epsilon >= 1 from the parameters()
calculation of the stress-strain relationship.

A problem with this is that now calibrate_theta_F.R in the
whalecollision project now creates a strange left-hand graph (the one
showing force misfit as function of theta). So, the next step is to look
into what's going on there