remove trial method for altering a.
This was going to be too confusing to the user.  I realize that there's no need for it, because the app can have a slider, and I can make a diagram for the paper by plotting Vcrit vs E, with a line for large and small ships and a 'standard' whale.