Test that solution is same as in 2020
This adds a dataset, "sol20200708", that is the output of strike() as
the package existed on July 8, 2020, prior to the publication of the MMS
paper. The main purpose is for a new test that assures that the key
functions of the package have not been inadvertantly changed in a
backwards-incompatible way.

Kelley, Dan E., James P. Vlasic, and Sean W. Brillant. “Assessing the
Lethality of Ship Strikes on Whales Using Simple Biophysical Models.”
Marine Mammal Science, October 12, 2020, mms.12745.
(https://doi.org/10.1111/mms.12745).